requirements.txt
Flask==2.3.2 
opencv-python==4.8.0.76 
numpy==1.24.3 
Pillow==9.5.0